Override nested type name for `gcp.kms.Registry.eventNotificationConfigs`
The nested type for this new property conflicts with the nested type of the existing (now deprecated) `eventNotificationConfig` (singular) property. The conflict happens because the new property is an array, and we singularize the nested type name for arrays (which conflicts with the existing singular property). To fix, we provide a manual override of the name, appending an `Item` suffix to the new nested type.